replace travis with circleCI to run full tests (#84)
* Add .circleci/config.yml

* specify working directory

* update run command

* comment out run command to see if it will build

* uncomment run & clean up'
:

* fix formatting

* change go version to 1.14

* add make vet

* fix indentation

* scrap travis

* don't run this test in parallel, see if it works

* that wasn't it

* add a 200ms sleep after creating custom roles to see if it fixes the not found error

* limit to 3 packages at a time

* try running just one package at a time?

* try a sleep of 1 second after custom role creation

* only clean team members that are pending invite

* run only team member tests

* run only create test

* only run update with custom role

* try all the team member tests again with seat limit flag off

* run all team member tests separately

* add test clean account

* remove -p and add custom props

* through test acc access token

* fix test name typo

* add all tests separately

* remove sleep

* don't run cleaner test

* run creates separately

* forgot to commit test name change

* run all four tests separately

+++ launchdarkly/account_cleaner_test.go
@@ -87,7 +87,7 @@ func (c *Client) cleanTeamMembers() error {
 		return handleLdapiErr(err)
 	}
 	for _, m := range members.Items {
-		if *m.Role != ldapi.OWNER_Role {
+		if *m.Role != ldapi.OWNER_Role && m.PendingInvite == true {
 			_, err := c.ld.TeamMembersApi.DeleteMember(c.ctx, m.Id)
 			if err != nil {
 				return handleLdapiErr(err)
